agentception/pyproject.toml
[project]
name = "agentception"
version = "0.1.0"
description = "Multi-tier AI agent pipeline dashboard"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
    "aiofiles>=23.2.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-anyio>=0.0.0",
    "mypy>=1.8.0",
    "types-aiofiles>=23.2.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
agentception = "agentception.app:main"

[project.urls]
Repository = "https://github.com/cgcardona/maestro"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentception*"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true
warn_unreachable = true
show_error_codes = true

# Third-party stubs are not always available; suppress missing-import noise
# for libraries that ship without inline types.
[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "alembic.*", "asyncpg.*", "aiosqlite.*"]
ignore_missing_imports = true

# pytest decorators are untyped by design; relaxing this keeps test files clean.
[[tool.mypy.overrides]]
module = ["agentception.tests.*"]
disallow_untyped_decorators = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    # asyncio subprocess transports occasionally raise RuntimeError("Event loop
    # is closed") in their __del__ finaliser when the test event loop tears down
    # before GC has collected all transport objects.  This is a CPython /
    # asyncio implementation detail that is harmless in tests — the subprocess
    # already completed; the cleanup failure has no effect on test correctness.
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
markers = [
    "e2e: end-to-end smoke tests that exercise the full workflow with mocked GitHub",
]
